separate symbol table interface from implementation
