+++ src/include/souffle/SouffleFunctor.h
@@ -1,0 +1,21 @@
+/*
+ * Souffle - A Datalog Compiler
+ * Copyright (c) 2022, The Souffle Developers. All rights reserved
+ * Licensed under the Universal Permissive License v 1.0 as shown at:
+ * - https://opensource.org/licenses/UPL
+ * - <souffle root>/licenses/SOUFFLE-UPL.txt
+ */
+
+/************************************************************************
+ *
+ * @file SouffleFunctor.h
+ *
+ * Main include file for Souffle C++ functors
+ *
+ ***********************************************************************/
+
+#pragma once
+
+#include "souffle/RamTypes.h"
+#include "souffle/RecordTable.h"
+#include "souffle/SymbolTable.h"
